# api/openapi.yaml
openapi: 3.1.0
info:
  title: MailRules API
  version: 0.9.0
  description: |
    JSON REST API of the MailRules daemon, plus one Server-Sent Events stream. This file is
    the contract the web UI's TypeScript client is generated from. A test fails if the daemon
    serves an /api path that is not listed here, or the reverse.

    **Conventions**

    - JSON bodies. Unknown fields in a request are rejected with 400 `invalid_json`.
    - Timestamps are unix seconds. Ids are integers. An absent time or reference is `null`.
    - Errors are `{"error": {"code", "message", "path"}}`. `code` is stable and machine-readable;
      `message` is a sentence safe to show; `path` names the field when there is one
      (for a rule: `conditions.all[0].op`).
    - Lists that can grow use cursor pagination: `?cursor=&limit=` (limit 1 to 100, default 50).
      The response carries `next_cursor`; pass it back as `cursor`, and `null` means the end.
    - Every route except the four under `/api/auth`, `/healthz`, `/readyz` and `/metrics` needs
      the `mailrules_session` cookie, and answers 401 without one: `setup_required` before the
      admin account exists, `unauthenticated` after.
    - Every response hands out a `mailrules_csrf` cookie, and every non-GET request must send
      its value back in the `X-CSRF-Token` header, or it is answered 403 `csrf_failed`.
      Call `GET /api/auth/me` first: it hands out the cookie and says which screen to show.
...
servers:
...
security:
  - session: []
...
paths:
...
  /api/actions/undo:
    parameters:
      - { $ref: '#/components/parameters/Since' }
    post:
      tags: [Fixes]
      operationId: undoSince
      summary: Undo everything done since a time ("Undo the last hour")
      description: Recorded as one batch of kind `undo`. An action that cannot be undone does not stop the rest. The Needs review tag is not an undoable action and is left alone.
      responses:
        '200':
          description: The undo batch and its counts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UndoResult' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: since)
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
...
  /api/batches/{id}/undo:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Fixes]
      operationId: undoBatch
      summary: Undo every action of a batch, newest first
      description: An action that cannot be undone does not stop the rest; `failed` counts them and the batch becomes `undone` only when every action was.
      responses:
        '200':
          description: The batch and how the undo went
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UndoResult' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: mailrules_session
      description: Set by `/api/auth/setup` and `/api/auth/login`. Non-GET requests also need the `X-CSRF-Token` header.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    Since:
      name: since
      in: query
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
      description: Unix seconds; everything done at or after this time is undone
  responses:
...
    Invalid:
      description: The request is wrong; `code` says how and `path` names the field
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Unauthenticated:
      description: No session. `setup_required` (show first-run setup) or `unauthenticated` (show sign-in)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    CsrfFailed:
      description: '`csrf_failed`: missing or wrong X-CSRF-Token header'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    NotFound:
      description: '`not_found`'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
...
  schemas:
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code, e.g. invalid_input, csrf_failed, unauthenticated
            message: { type: string, description: Sentence safe to show the user }
            path: { type: string, description: Field the error is about, when there is one }
...
    Batch:
      type: object
      required: [id, kind, status, total, done, created_at, actions, account_id, folder, since, tokens, cost_usd]
      properties:
        id: { type: integer, format: int64 }
        kind: { type: string, enum: [live, cleanup, review, correction, undo], description: '`live` is one per UTC day of live processing, so "undo today" is one batch undo' }
        status: { type: string, enum: [running, done, failed, undone], description: 'A cleanup is `running` until it has been through its emails, then `done`; `failed` when it was cut short (what it did stays, and can be undone)' }
        total: { type: [integer, 'null'], description: 'How many items it set out to handle; null when not counted up front (live). For a cleanup, the emails selected' }
        done: { type: integer, description: 'How many it has handled; for an undo batch, how many actions were undone' }
        created_at: { type: integer, format: int64 }
        account_id: { type: [integer, 'null'], format: int64, description: 'Cleanup: the mailbox it sorted. null for the other kinds, and once that account is deleted' }
        folder: { type: string, description: 'Cleanup: the folder it sorted. Empty for the other kinds' }
        since: { type: [integer, 'null'], format: int64, description: 'Cleanup: only mail received from this time on; null = all of it' }
        tokens: { type: integer, description: 'Cleanup: model tokens used so far, in and out. 0 for the other kinds' }
        cost_usd: { type: number, description: 'Cleanup: model cost so far. 0 for the other kinds' }
        actions:
          type: object
          description: How many of the batch's own actions have each status. An undo batch has none of its own
          required: [done, dry_run, failed, undone]
          properties:
            done: { type: integer }
            dry_run: { type: integer }
            failed: { type: integer }
            undone: { type: integer }
    UndoResult:
      type: object
      required: [batch, undone, failed]
      properties:
        batch: { $ref: '#/components/schemas/Batch' }
        undone: { type: integer, description: Actions undone by this call }
        failed: { type: integer, description: 'Actions that could not be undone: the email is gone, or its account is not connected' }
...